MB-42204: Disallow schedule() during unregisterTaskable()
+Problem+

When using FollyExecutorPool there is an intermittent hang in CV tests
during bucket shutdown. This results in the test suite timing out.

The hang occurs when a bucket is being shutdown and we attempt
to schedule a new task via ExecutorPool::schedule().

The problem is caused by the way FollyExecutorPool::unregisterTaskable
is implemented:

1. Mark all scheduled tasks as cancelled.

2. Waits for all tasks to complete cancellation (be removed from
   taskOwners).

However, if another, new task is attempted to be scheduled between (1)
and (2) then it effectively "misses" the cancellation, and (2) will
wait forever.

+Solution+

Add an additonal 'registered' flag to each registered Taskable in
taskOwners, initially set to true when a Taskable is registered. This
flag is checked in schedule(), if false then the task is not
scheduled.

At the start of unregisterTaskable this flag is set to false; ensuring
no more tasks can be scheduled once Taskable unregister has begun.

Change-Id: I8cc24a9b4546b72a71b6ac19953df8dedcaf2dd7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138576
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>