MB-41747: Reset the cached couch handle
With the introduction of a cache containing the vbucket files
we would reuse the couchstore handle every time we tried to
write to the database. That would work in the case where no
errors occur, but in the case for commit failures the data
was already written to the disk and the internal datastructures
in the couchstore handle referenced the new data (and it would
be included in the next commit).

On flush failures we would try to "roll back" the changes we
did and previously this worked due to the fact that we would
"drop" the couchstore instance every time we would try to add
data to the instance.

This patch resets the database handle every time it is fetched
from the cache. Ideally it should only be reset on failures
which would require it to do so, but that should be tracked in
its own MB.

Change-Id: I5f4809d9b5e266bb02a3c011e7316162201bab54
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138740
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>