MB-41321: Only delete _local/|cid| if the collection is dropped
Development of MB-41882 observed some issues with a local document
being written multiple times in one-commit, there is no guarantee
which version will end up being committed.

It was observed that the 'Flush::droppedCollection' container was
being processed and resulting in a delete of the collection stats
with no check to see if the collection had actually been
re-created (and even had items stored). Inspecting some flush
batches for the "resurrection" tests showed a flush with both
a write and delete of the same local document.

This change only deletes the stats when the flush batch drops a
collection as the final (most recent) event for the collection.

Change-Id: I55799e2ba02089352d458b1177142fda5144b699
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138877
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>