Verify that we don't underflow the on_disk_prepares
We've seen a number of problems where we've purged more
prepares than we expected resulting in an underflow of a
counter causing the system to crash whenever we tried
to _read_ the value off disk (which we normally don't
do as we've got a cached version in memory). This patch
cause us to crash at the time when we _write_ the
illegal value to disk.

Also corrected the method name in log messages to make
it easier to identify where the log message comes from.

Change-Id: Id35c5046bb4c607ed4676d1838dd473f798b9284
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138908
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>