MB-38428: Run flusher and compactor in parallel
Release the exclusive lock for the vbucket while doing the
full compaction of the vbucket and then run 10 attempts
to catch up without holding the lock while replaying the
changes.

To make sure that we don't "race" forever trying to catch up
we'll hold the lock the last time we try to catch up.

There has been a number of bugs where the number of prepares
found on disk differs from the number in _local/vbstate so
this patch also includes a check which scans the on disk prepares
and compares the number with whats stored in _local/vbstate.
That piece of code is to be deleted once we've verified that
everything is as expected.

Change-Id: I28e9d69764feba6d78be4208d9a06408d9e26f92
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139408
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>