MB-41430: Destroy magma instances before we call dtors at destruction
During bucket destruction we currently destroy magma instances in the
destructors of the MagmaKVStores which are destroyed by the KVShards
and so on up the chain to the KVBucket instance. Magma defers the
destruction of their background tasks (compaction is the one we care
about here) to the destruction of the magma instance. As the compaction
task can call back into the EPBucket to make the compaction context
we need to ensure that this cannot happen whilst we destroy the
EPBucket. Do this by moving the magma instance destruction out to
the deinitialize function of EPBucket.

Change-Id: Ief66c91ff82885b7175fbd3e107a48cfb9d0347e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/138962
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>