MB-42309: Setting stream to dead not working for stream-id streams
The setStreamDeadStatus was geared up for finding a stream with
a matching ID but no caller ever passes an ID. The function is
actually incorrect and all callers will never pass an ID (vbucket
rollback or state changed).

For all uses of this function we need to unconditionally end every
stream associated with the producer.

Change-Id: I997a4b1ccaae3661edd0ca4e16b826a46a904257
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>