MB-42282 magma-kvstore: Use disk snapshot for bySeqno scan
The bySeqno iterator requires the snapshot to be held for the duration
of the scan. The default snapshot holds inmemory memtables at that
point-in-time as part of the snapshot. This results in signficant memory usage
if the snapshot needs to be held for a longer duration.

This issue can be resolved by forcing writecache flush for the
kvstore and obtain the diskonly snapshot for performing bySeqno scan.

Change-Id: If7c4c6028509a3ab97acfb79fccf6f169b3def81
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139705
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>