MB-40934: Prevent DcpProducer disconnection on Status::KeyEnoent
Currently the DcpProducer will disconnect due to a seeing a
cb::mcbp::Status::KeyEexists or cb::mcbp::Status::KeyEnoent for a
DcpPrepare, DcpCommit and DcpAbort. As in the case KeyEnoent is send it
could be due to a situation where DcpCommit or DcpAbort has failed due
to not finding an prepare. However, cb::mcbp::Status::KeyEnoent can be
returned due to the DcpConsumer not finding a stream for the DCP
message's vbucket.

To fix this going forward, we will introduce two new status codes,
DcpStreamNotFound and OpaqueNoMatch to clear up the difference in use
between KeyEnoent use for not finding a prepare and not finding a
stream. OpaqueNoMatch to clear up the difference between KeyEexists for
a miss matching opaque or for a situation where the engine has returned
cb::mcbp::Status::KeyEexists.

To enable these new messages we will introduce a key
"v7_dcp_status_codes" with a value of "true" with a DCP_CONTROL
message.

This patch also adds tests to ensure this works and updates unit test
that verify DcpProducer::handleResponse() behaviour.

Change-Id: I64c18e4126f64d8a27db1e0715922221d63f8e04
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139602
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>