MB-41882: Detect an empty collection being dropped and skip purge
Dropping a collection on a persistent bucket results in triggering of
compaction so that KV can reclaim *all* resources consumed by the
collection, that is memory and disk used by items. It is desirable to
avoid triggering compaction and there is no need if the collection has
never been written to by a committed item. If a collection has had
zero committed items we can detect this by comparing the collection's
high-persisted seqno against the collection's start-seqno when equal
the collection is 'empty' and does not need a compaction driven purge.
A collection which empty and was created and dropped in a single flush
is also detected, this is a collection with 0 as the high-seqno.

Change-Id: Iba1e6840c8d157fe30fed3183f381743f7723b85
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/137968
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>