Avoid recreating MetricServer when setting same config
If the same config is set repeatedly, normally a new MetricServer will
not be created because the config has not changed.

However, when setting the port to 0 this was not previously the case.
Instead, the config would be updated with the allocated port the
MetricServer was actually listening on. As a result, setting the
original config (with port=0) again _did_ recreate the MetricServer, as
the config no longer matched the original.

Change-Id: I78ed3a56859a6de11d12ba46e1c573f35265a3df
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139596
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>