Fix DcpConsumer::deletion() status at opaque miss match
When the opaque for a message does not match the streams opaque in
DcpConsumer::deletion() return ENGINE_KEY_EEXISTS in the same manner as
DcpConsumer::lookupStreamAndDispatchMessage(). Rather than using
ENGINE_KEY_ENOENT, which is also used to return a status when a stream
has not been found for the messages vbucket.

This issue was found when investigating MB-40934.

Change-Id: Ife83553dd6291e1d1c64ef0db2b169cc6cb20629
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139606
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>