Reset send queue size before returning to libevent
We currently check if a connection is stuck transmitting data
to the client before starting to execute commands and record
the current transmit queue and the current time.

If the connection _isn't_ stuck, but have data in its output
buffer we might add more data to the output buffer and the
kernel may accept exactly this amount of data leaving the
queue size the same causing us to think the connection is
stuck and shut down the connection.

This patch checks (and possibly adjust the send queue size and
timer) after processing all cookies.

Change-Id: I2e0d872ee8f5ceb837dc66fec0e7bbd88b475d4a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139643
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>