MB-41804: Disable test_mb19982
This test flips a vbucket replica->active->replica and manually creates
snapshots and mutations with pre-determined seqnos.

This fails under ephemeral if the vbucket attempts to page items while
it is briefly active; removing old data to reclaim memory creates
deletions, which advance the current high seqno. The test does not
account for this, and so sends snapshots and mutations at invalid
seqnos.

This test should be investigated and re-enabled once this is resolved.

Change-Id: I5d51332298d76325b1e5908452c9db48e27f1772
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140017
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>