[Refactor] Make ItemIface a real class instead of void
With this change the ItemIface is a base class the different
engines may subclass for their own implementation.

default_engine stores all of its internal objects as hash_item
and we don't want to increase the size of all of them with 8 bytes
so instead a new ItemHolder class is introduced which is the one
being used between the core and default-engine.

ep-engine already had an Item class which is used between the
core and the engine (which keeps all items in the cache as
StoredValue objects) so this class just inherit from ItemIface.

Change-Id: Ibe8afd3903c687369f65f2b630c3e0f31f82f610
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/139882
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>