[BP] MB-42610: Do not expire prepares when committed items exist
Currently if a committed item exists then it's possible for the
expiry code to attempt to expire a prepare. This obviously breaks
durability and it can lead to us having two committed items in the
HashTable for the same key which will throw assertions the next time
we lookup that key.

This can only happen in a very specific and limited set of
circumstances. The vBucket in question must go down having only
partially persisted a Disk snapshot. The vBucket must then flip from
replica to active (a situation that would lead to data loss) then
we must attempt to expire an item (via compaction or the pager).
This expiration could find a prepare in the HashTable and attempt to
expire it if the item has the same cas (this should only be possible
when we load a completed prepare).

Fix this issue by skipping expiration if we find a prepare with the
same cas of our item in the HashTable.

Change-Id: I1141b0090b1ff65eabf611da1b478b72bd284c78
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142089
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>