MB-42592: Add purge age to tombstones in GetExpiryTime for magma
Magma calls an API function that KV supplies to get the expiry
time of an item. Magma uses this API to generate expiry time for
histograms that determine when compaction runs. This API is also
used for tombstones to determine when they should be compacted.
This is problematic for magma as the time returned does not
correspond to the time at which compaction would ideally run.

Modify the GetExpiryTime API in KV such that it adds the purge time
to the deleted times of tombstones.

Change-Id: If3406e829da55ecf6666442ce3035d2bebe1ec53
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140139
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>