MB-39946: Don't record the system event size for collections
MB-39946 is being difficult to track down, it seems that an
underflow is required in the disk-size, but this only happens
for the default collection because when empty, that collection
should have a disk size of 0.

Other collections whilst empty have a disk size of around 60 bytes
(the system event) so they may be hiding more instances of the
under flow - skipping recording of the system event should help
increase the chances of hitting this MB in more scenarios.

Change-Id: Ic68bd777037dc99cb1dd8a6359f00ccd1b5ce356
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140211
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>