MB-42608: Add test for http://review.couchbase.org/c/kv_engine/+/109841
http://review.couchbase.org/c/kv_engine/+/109841 introduced in MB-33332.

Purpose of the test is showing that we cannot mark-stale a Committed
Prepare unconditionally at EphemeralVBucket::commitStoredValue.

The reason is that the following may happen:
  1) Another prepare is queued again for the same key
  2) A RangeRead covers both the first and the second prepare
  3) At commit we mark stale the second prepare

At this point the RangeRead covers Prepare:1, Commit:2 and Prepare:3.
With both Prepare:1 and Prepare:3 stale OSVs with no replacement (see
Ephemeral replacement / seq-list-dedup for details).
A backfill snapshot will be sent to replica with duplicate key at
seqno:1 and seqno:3.

Change-Id: I42c318e580c79bd2311ed975cf6d523b4217982e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140369
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>