MB-42083: Don't dereference next if end in PDM
In some failure cases we can end up with next in PDM::completeSyncWrite
pointing to end. Currently this can segfault if enforceOrderedCompletion
is true due to the dropped collections skipping code. Correct this so
that we skip this case if next points to end and throw the exception in
the following if block instead.

Change-Id: Idb04c8aab61f9ae5e0fdc92603e92be5d84fc207
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140352
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>