[BP] MB-42918: Do not replace committed item with pending at add
Currently when we do an add we can replace certain committed items
with pending ones. One of these cases is a when we have an unpersisted
delete. If we replace this with a pending StoredValue then we allow
a subsequent get to find no committed StoredValue in the HashTable and
potentially do a BgFetch. This BgFetch may return the alive document
back to the HashTable/client which it should not do.

Fix this by creating a new StoredValue if we would have replace a
committed one with a prepared one.

Change-Id: Ide30894867ec9710ab0d334a97acd5c6c8d009ff
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142093
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>