MB-42918: Do not replace committed item with pending at add
Currently when we do an add we can replace certain committed items
with pending ones. One of these cases is a when we have an unpersisted
delete. If we replace this with a pending StoredValue then we allow
a subsequent get to find no committed StoredValue in the HashTable and
potentially do a BgFetch. This BgFetch may return the alive document
back to the HashTable/client which it should not do.

Fix this by creating a new StoredValue if we would have replace a
committed one with a prepared one.

Change-Id: I85bb2b249afa8894c594d9d4eaed056a7fe7e833
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140592
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>