MB-42780: Make replica resilient to missing MARKER_FLAG_CHK
Since 6.5.0 the Active ensures that Disk snapshots are never merged into
an existing checkpoint. That is achieve by setting the MARKER_FLAG_CHK
in the SnapshotMarker sent to Replica.

That all works fine as long as Active behaves as expected.
In pre-6.5.0 we had issues where Active could miss the MARKER_FLAG_CHK
in markers. In the 6.0.1->6.6.1 upgrade scenario seen in MB-42780, one
of the effect is replica failure (with consequent rebalance/upgrade
failure).

With this patch, replica enforces the same logic that we had already
added at active in 6.5.0. That makes replica resilient to any DCP
Producer that misses to set the MARKER_FLAG_CHK properly.

Change-Id: I9b32dec1caa5b262f0cb524317624e4e955481b2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140626
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>