MB-42861: Avoid trying to format unsanitized stat keys
Some cbstat keys contain uncontrolled data, for example connection
stats contain the connection name, which is set by the remote end.

This data is not guaranteed to be free from invalid fmt specifiers,
thus the following is likely to throw

 fmt::format("ep_dcpq_some_{weird}_connection_name:type", ...);

This only occurs for stats which have their key manually composed and
added through add_casted_stat. Stats which are added through the
collector will specify variable parts of the key as labels,
essentially leading to:

 fmt::format("{connection_name}:type",
    fmt::arg("connection_name",
             "ep_dcpq_some_{weird}_connection_name"));

This is perfectly fine, and the provided connection name appears in the
stat key as intended, regardless of any "{...}".

Change-Id: If395985f252f44d044c7e17ab7d82f2e059c75fa
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140689
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>