MB-39946: saveDocsCallback newInfo must not be null
couchstore would only set this null for a failure of cb_malloc
or possible corruption issue (where an internal size became -ve).

If newInfo was null and we silently carried on, that could be
a cause of MB-39946.

Note there is no evidence to suggest this is the cause of MB-39946.

Change-Id: Icf031a26f9b5721408b236c18e08045ce7394b2c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140698
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>