MB-39292: 4/n Detect change of immutable property
When updating vbuckets to new/current manifest, compare
all properties of the scope/collection. If for example
the name is not equal use the new
'ImmutablePropertyModified' status to flag such a case.

In the case of a force=true manifest, react to the status
by recreating the collection (with new properties).

Change-Id: I158dfd42cde1112f795d8d9c33574c4c3a5a81c3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144494
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>