MB-39505: Copy AddStatFn into CBStatCollector
Keeping a const ref member variable may encounter lifetime issues -
for example:

 static void append_stat(std::string_view key,
                         std::string_view value,
                         gsl::not_null<const void*> void_cookie) {
 }
 ...
 CBStatCollector(append_stat, cookie);

This leads to a stack-use-after-scope as the AddStatFn member is then a
const ref to a _temporary_ std::function created from the function ptr.

A CBStatCollector is only expected to be created once per stat request,
so the cost of the copy is not excessive.

Change-Id: Ibebd5377df6795c1051cd26ff22d9bfb7747e625
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140888
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>