MB-42885: KVStore::rollback must pass 'logically dropped' items to the callback
The implementation of rollback uses a scan to iterate through the
documents that are "going away", both CouchKVStore and MagmaKVStore use
code which means any documents included in that scan belong to a dropped
collection are not passed to the callback.

The following example demonstrates the bug which can leave items in
the hash-table.

    5 6 7 8 9
    B A A A Drop(A)

In the above seqnos 6, 7 and 8 are items of collection A which was
dropped at seqno 9, no purge has occurred yet.

If rollback chose seqno 5 as the rollback point it will issue a scan
from seqno 6 to 9 and KV-engine uses the items passed to the
'rollback callback' to clean/fix the hash-table (and item count). In
this example 6, 7 and 8 will not be passed to the callback, leaving
those items in the hash-table and an incorrect item count.

Change-Id: Ia94772d464d6120153fb3ba38c37ec6c2f0aa356
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140897
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>