[Refactor] Store the DCP conn handler in its own variable
ep-engine used to store the "conn handler" object which represents
the DCP object in the "engine-specific" part of the cookie, but
that field is also used by other commands for different things
(a sync-write would use it to keep the cas; compaction use it
to store a setting that it is running etc). If any of these
commands would be received on a DCP connection they would
clear the engine-specific part in the cookie when they
complete and none of the methods in the DcpIface would work.

This patch change that logic to store the conn handler in
the Connection object.

Change-Id: I9850b1b30881868a9ba1731cea387f1014368fb9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/140956
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>