MB-41942: Change PCS/PPS GSL asserts to exceptions
Change PCS and PPS GSL assertions in EPBucket::flushVBucket to
normal exceptions so that we can log some extra information
about the issue (the checkpoint type that we are currently
processing). This will help us narrow down the code paths that
may be taken in this function when we are attempting to update
the value.

Change-Id: I58ab0b8c61554aa151c91022d6d740f9d5d2902c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141052
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>