MB-42976: Skip test_multi_set of ep_testsuite_basic for Magma
test_multi_set() of ep_testsuite_basic.cc keeps timing out for magma
under TSAN, this is likely to be due to a performance issue with Magma.
So skip the test for Magma to ensure we don't block patches in CV, while
the issue is investigated by the Magma team.

Change-Id: I179b3680858fa3d70e73ca1a3614e91f73c33e6b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141132
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>