MB-41942: Persist the VB state earlier during warmup
Currently we attempt to persist the vBucket state(s) when we complete
warmup. This leads to a window during warmup in which we can write
new items (via expirations) after identifying that we need to generate
a new FailoverTable entry but before we persist it. If we then crash
or shut down in this period then we can fail to roll back any replica
nodes. This can in turn cause data loss/data corruption.

Correct this by persisting the vBucket state during warmup as soon as
we know that we need to generate a new FailoverTable entry.

Change-Id: Iea5eb60b29ae4ba1d66e83d5b5e4f2b44d516c26
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/142080
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>