MB-42980: Fix failures in test_dcp_consumer_flow_control_aggressive
Steps in the test are:

1) Create some connections and check the expected DCP buffer size
2) Remove some of them
3) Wait for num-dead-conns==0 for ensuring (2) completed
4) Check the new expected DCP buffer size for each remaining conn

Problem is in (3). Connections are released (and num-dead-conns
decremented) in DcpConnMap::manageConnections, where we do:

a) Acquire DcpConnMap::connsLock
b) Decrement num-dead-conns
c) Release DcpConnMap::connsLock
d) Release dead conns -> This is where we reset the DCP buffer size for
   all the remaining connections

The problem is that (3) can be satisfied after (c) but before (d). Which
means that we may end up doing the checks at (4) before (d) has
completed. Thus the intermittent failures.

Change-Id: I880f06bda40f53abd4796a13f1f7c2ac4b0427e2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141228
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>