MB-43129: Allow reconfig of interfaces having same host and port
Going from [1] as below,
[
 {
    "port": 12000,
    "host": "*",
    "ipv4": "required",
    "ipv6": "off"
 }
]

To [2]:
[
 {
    "port": 12000,
    "host": "*",
    "ipv4": "off",
    "ipv6": "optional"
 },
 {
    "port": 12000,
    "host": "*",
    "ipv4": "required",
    "ipv6": "off"
 }
]

The diff logic prior to this change would drop an IPv4 interface whose
host is <<"*">> and port is 12000, even though the new wanted interfaces
where as per [2] requiring IPv4. The problem is that the current logic
would decide to drop the IPv4 interface after comparing with the 1st
optional interface which is IPv6.

Change-Id: Ie386aa4124dd2e38a4d7f41d973aef7237eabaa8
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141254
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>