Don't run all tests over SSL and PLAIN
A lot of our test cases tests the actual implementation of a given
command (and not the actual "state machinery for the network layer".
Previously we had our own "implementation" of sending and receiving
data over SSL (with our own buffering before doing send / receive
on the actual socket) and wanted to make sure that we ran as much
testing as possible of the state machinery modification that
introduced. As part of moving to buffereven all this code is gone
and the only difference in our code between SSL and !SSL is related
to accepting the connection (where SSL have some extra steps).

Due to this we can stop running all of these tests on both PLAIN
and SSL (and only run them over SSL). This should significantly
reduce the number of sockets being created as part of running
our tests (see MB-30022) and speed up test execution as we'll
only run the tests once. Note we've got some test cases which
already only ran over PLAIN so we'll still test that code path.

Change-Id: I3e17cd2136b44541d7d555700c8e34ec39ac200b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141330
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>