MB-41300: Reposition the HPS correctly in PDM at Prepare deduplication
When we have a PDM::trackedWrites like

  P(keyA):1    P(keyB):2
               ^
               HPS:2

replica may receive a duplicates for some existing Prepare in a Disk
snapshot, eg another Prepare for keyB, and we erase the previous
Prepare. If that happens, currently we wrongly reset the HPS::it to
trackedWrites::end:

  P(keyA):1    x    P(keyB):3
               ^
               HPS:2

This is an invalid state and a pre-condition for a crash.
As soon as we try to move the HPS, we set HPS::it to
trackedWrites::begin and we try to set HPS::lastWriteSeqno to 1, which
breaks the monotonicity invariant on lastWriteSeqno and throws.

When we erase a Prepare by deduplication, HPS must be repositioned to
std::prev(HPS) in general. HPS::it=trackedWrites::end is allowed only if
trackedWrites is empty.

Change-Id: Ifa516e771e422fdeb175e81e8dad163f17702436
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141418
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>