MB-41165: Rename StatisticsUpdate persistedHighSeqno
To be able to add the size of prepares to the disk size of the
collection we must modify the stats tracking code to allow us to
add stats when we have a collection with only prepares in it.
Currently we have many persistedHighSeqno checks that we
cannot pass so no stat doc gets written. To fix these we need to track
a high seqno that includes prepares. To follow the same naming
convention as vBucket level stats, rename persistedHighSeqno to
persistedCommittedSeqno. In a future patch I'll add a new
persistedPrepareSeqno.

Change-Id: Ife81d820c5f1a58931ee23e5ff8b4b103603dec1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141482
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>