MB-41092,MB-43081: Ensure saveDroppedCollection() stores HighSeqno
Currently Manifest::saveDroppedCollection() stores the persisted on
disk high seqno for the dropped collection. This can cause an issue
where we don't trigger compaction after the drop collection. Causing
on disk item count to look wrong which the user will see in full
eviction. This is due to Flush::countNonEmptyDroppedCollections()
which decides if there are any non empty collections that have been
dropped in a flush. This function then performs the check
highSeqno != 0 && highSeqno != dropped.startSeqno. Which will be false
if the persisted seqno is 0. This can happen when the collection drop
system event is added to the dirty queue before any mutations have
been persisted to disk.

To fix this we change Manifest::saveDroppedCollection() to store the
collections high seqno over the persisted high seqno in the
Manifest::droppedCollections map.

This fix also addresses MB-43081 as it ensures we schedule compaction in
a situation where the collection only contained prepares.

Change-Id: I14ea5a5d9fefaaf3f3ed62653ceef2ac87f28e15
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>