MB-42795: Replace setup/reset context with switch execute env
In the old scheme of hiding the ep engine context in the setup routine,
we are unable to track back to the setup/reset functions easily and
lose context when utilizing magma threads for sstable flushing and
compaction. The new scheme will take 3 arguments:

- SwitchEnv function that allows magma to switch to a specific context
  and the function will pass back the previous context.
- ExecutionEnv is a pointer to a specific execution environment
- GlobalEnv is a pointer to the default execution environment

We treat a null execution environment ptr as global but in the future
if a specific global environment is needed, the GlobalEnv exists. It
also helps identify/document when we switch to the default environment.

A compactor thread picks up a compaction task and sets the environment
to the kvstore being compacted. During compaction, we need to create
and remove sstables which require accessing the global fd cache. All
allocations to the global fd cache need to be accounted for in the
default/global environment. The code path in magma might encounter more
than switch to the global environment.

Example)
- SwitchEnv to user defined (bucket) environment
- switchEnv(Global)Guard
- switchEnv(Global)Guard
- destructor of 2nd guard
- destructor of 1st guard
- reset back to original environment

Change-Id: I0afbfcdeae0a2eef99c5ca229244d7175f10af71
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/141651
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>